ALGOR-14: migrate to the new Tinkoff Invest API

+++ src/main/scala/com/github/ppotseluev/algorate/core/Bar.scala
@@ -10,7 +10,7 @@ case class Bar(
     closePrice: Price,
     lowPrice: Price,
     highPrice: Price,
-    volume: Int,
+    volume: Long,
     endTime: OffsetDateTime,
     duration: FiniteDuration
 )
